requirements.txt
flask
python-dotenv
google-genai
duckduckgo-search
requests